Initial commit
Ok this is the worst initial commit I could possibly do, mostly because
this project has literally been going for years.  This is the
culmination of a few years worth of poking around and building things.
It's progressed from standalone sensors attached to a board to being on
the board, different chips, etc.

I'm using the max31856 design (prod-onboard) right now.  Channel 4 is
sorta giving me trouble staying online, but it's not clear why.
Otherwise I've been very happy with the board overall.